Fixed some tests
